% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_practical.R
\name{causalcode}
\alias{causalcode}
\title{Render Code}
\usage{
causalcode(part = 1, dir = getwd())
}
\arguments{
\item{part}{choose from 1,2,3}

\item{dir}{directory, default is getwd()}
}
\value{
an HTML file
}
\description{
Render Code
}
\examples{
\dontrun{
causalcode(1)

# where is the output html-file?
getwd()
}
}
