% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CItests.R
\name{GCM}
\alias{GCM}
\title{Wrapper function for the 'Randomized Conditional Independence Test (RCIT)' from the 
comets package}
\usage{
GCM(x, y, S, suffStat)
}
\arguments{
\item{x}{(integer) position of variable X}

\item{y}{(integer) position of variable Y}

\item{S}{(integer) position of all variables in the set S}

\item{suffStat}{a list with the element 'data'}
}
\value{
a p-value
}
\description{
Wrapper function for the 'Randomized Conditional Independence Test (RCIT)' from the 
comets package
}
