% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPDF.R
\name{openCode}
\alias{openCode}
\title{Open Code}
\usage{
openCode(file)
}
\arguments{
\item{file}{string.}
}
\value{
an HTML
}
\description{
Open Code
}
\examples{
\dontrun{
library(DataTrainCausalLearning)
openCode("Causal1_code.html")
openCode("Causal2_code.html")
openCode("Causal3_code.html")
}
}
