% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPDF.R
\name{openPDF}
\alias{openPDF}
\title{Open Practicals}
\usage{
openPDF(file)
}
\arguments{
\item{file}{string.}
}
\value{
a PDF
}
\description{
Open Practicals
}
\examples{
\dontrun{
library(DataTrainCausalLearning)
openPDF("Practical_part1_2025.pdf")
openPDF("Practical_part2_2025.pdf")
openPDF("Practical_part3_2025.pdf")
}
}
