% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openPDF.R
\name{openSolution}
\alias{openSolution}
\title{Open Solution}
\usage{
openSolution(file)
}
\arguments{
\item{file}{string.}
}
\value{
an HTML
}
\description{
Open Solution
}
\examples{
\dontrun{
library(DataTrainCausalLearning)
openSolution("xxx.html")
}
}
