% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_practical.R
\name{show_practicals}
\alias{show_practicals}
\title{Show Practicals}
\usage{
show_practicals(nr)
}
\value{
opens the folder with the practicals
}
\description{
Opens the folder with the practicals.
}
\examples{
\dontrun{
show_practicals()
}
}
