% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcgas}
\alias{tcgas}
\title{The Cancer Genome Atlas's Study of Head and Neck Squamous Cell Carcinoma}
\format{
A data frame with 362 observations on the following 8 variables:
\describe{
  \item{BAX}{a numeric or discretized vector}
  \item{CDKN2A}{a numeric or discretized vector}
  \item{CDKN1A}{a numeric or discretized vector}
  \item{HMGA2}{a numeric or discretized vector}
  \item{MDM2}{a numeric or discretized vector}
  \item{SERPINE1}{a numeric or discretized vector}
  \item{THBS1}{a numeric or discretized vector}
  \item{CDK6}{a numeric or discretized vector}
}
}
\source{
\url{https://portal.gdc.cancer.gov/projects/TCGA-HNSC}
}
\usage{
data(tcgas)

data(tcgadisc)
}
\description{
The Cancer Genome Atlas (TCGA, https://www.cancer.gov/tcga), a landmark cancer
genomics program, molecularly characterized over 20,000 primary cancer and
matched normal samples spanning 33 cancer types.
}
\references{
The Cancer Genome Atlas Network. Comprehensive genomic
characterization of head and neck squamous cell carcinomas. Nature 517, 576–582 (2015).
}
\keyword{datasets}
